remove showWorkPageChanges toggle
